rename and move store to state folder
